% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calc_error}
\alias{calc_error}
\title{Calculate Error for Errorbars}
\usage{
calc_error(x, fun.errorbar = "sd", na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector}

\item{fun.errorbar}{Either a character string specifying the method to calculate the error ("sd", "se", or "ci"), or a function that returns a numeric value.}

\item{na.rm}{A logical value indicating whether NA values should be stripped before the computation proceeds. Defaults to FALSE}
}
\value{
The calculated error value of the input vector
}
\description{
This function calculates an error value (standard deviation, standard error, 95\% confidence interval, or a user-supplied function) for a numeric vector.
}
\examples{
calc_error(c(1, 2, 3, 4, 5))
calc_error(c(1, 2, 3, 4, 5, NA), fun.errorbar = "se", na.rm = TRUE)
calc_error(c(1, 2, 3, 4, 5), fun.errorbar = "ci")
calc_error(c(1, 2, 3, 4, 5), fun.errorbar = function(x) max(x) - min(x))
}
