% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calc_error}
\alias{calc_error}
\title{Calculate SD/SE for Errorbar}
\usage{
calc_error(x, fun.errorbar = "sd", na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector}

\item{fun.errorbar}{A character string or function indicating how to calculate
the error bar. Available character options are "sd" for standard deviation,
"se" for standard error, and "ci" for a 95\\% confidence interval. A custom
function can also be supplied.}

\item{na.rm}{A logical value indicating whether NA values should be stripped before the computation proceeds. Defaults to FALSE}
}
\value{
A numeric value representing the calculated error bar.
}
\description{
This function calculates either the standard deviation or the standard error of a numeric vector, depending on the specified method.
}
\examples{
calc_error(c(1, 2, 3, 4, 5))
calc_error(c(1, 2, 3, 4, 5, NA), fun.errorbar = "se", na.rm = TRUE)
calc_error(c(1, 2, 3, 4, 5), fun.errorbar = "ci")
calc_error(c(1, 2, 3, 4, 5), fun.errorbar = function(z) max(z) - min(z))
}
